Moved processing from main to ConstraintManager, ...
Moved processing out of main.pm into ConstraintManager via a new routine -
'execute' (which calls 'check_processes', which does the main work).
Progress toward intended final functionality: an infinite loop is now
executed (in 'execute'), where for each iteration a new process table is
obtained and checked and a 'sleep 5' is performed.  (The sleep value
will be configurable in the finished version, of course.)