Enhanced to allow dynamic config. changes.
A bit of restructuring to allow/implement re-parsing, if there's been a
change, the configuration files each time through the main loop (just
before check_processes is run) so that changes in the monitor's behavior
can be effected dynamically, without needing to kill/restart the
process.

A couple small cleanup-changes/improvements.